+++ package.json
@@ -12,12 +12,11 @@
     "next": "12.2.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-
     "react-icons": "^4.4.0",
-
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
-    "styled-components": "^5.3.5"
+    "styled-components": "^5.3.5",
+    "styled-reset": "^4.4.2"
   },
   "devDependencies": {
     "babel-plugin-styled-components": "^2.0.7",
