+++ src/components/CommunityMain/Main.js
@@ -7,34 +7,29 @@ import SectionDeal from './SectionDeal';
 import SectionPPL from './SectionPPL';
 import SectionReview from './SectionReview';
 import SectionExhibition from './SectionExhibition';
-import SectionPremium from './SecitonPremium';
+import SectionPremium from './SectionPremium';
 import SectionBest from './SectionBest';
-
-// import 'swiper/scss'
-// import 'swiper/scss/navigation'
-// import 'swiper/scss/pagination'
 
 
 const MainBlock = styled.div`
-  padding:30px 0;
-  width:1136px;
-  margin:0 auto;
-  
+  padding: 30px 0;
+  width: 1136px;
+  margin: 0 auto;
 `;
 
 function Main() {
   return (
     <MainBlock>
-      <SectionMain/>
-      <SectionNav/>
-      <SectionStory/>
-      <SectionCate/>
-      <SectionDeal/>
-      <SectionPPL/>
-      <SectionReview/>
-      <SectionExhibition/>
-      <SectionPremium/>
-      <SectionBest/>
+      <SectionMain />
+      <SectionNav />
+      <SectionStory />
+      <SectionCate />
+      <SectionDeal />
+      <SectionPPL />
+      <SectionReview />
+      <SectionExhibition />
+      <SectionPremium />
+      <SectionBest />
     </MainBlock>
   );
 }
