+++ src/components/CommunityMain/SectionCate.js
@@ -1,0 +1,175 @@
+import React from 'react';
+import Image from 'next/image';
+import { MdKeyboardArrowLeft, MdKeyboardArrowRight } from 'react-icons/md';
+import * as S from './main-styled';
+const SectionCate = () => {
+  return (
+    <S.Section height={82} cate>
+      <h2>카테고리별 상품 찾기</h2>
+      <div className="ul-area">
+        <ul>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate01.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>가구</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate02.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>패브릭</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate03.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>가전</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate04.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>유아·아동</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate05.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>조명</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate06.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>주방용품</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate07.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>데코·식물</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate08.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>수납·정리</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate09.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>생활용품</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate10.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>생필품</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate11.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>공구·DIY</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate12.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>인테리어시공</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate13.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>반려공물</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate14.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>캠핑용품</span>
+          </li>
+          <li>
+            <span>
+              <Image
+                src="/community-main/cate16.jpg"
+                layout="fill"
+                alt="cate"
+              />
+            </span>
+            <span>실내운동</span>
+          </li>
+        </ul>
+        <div className="btn-area">
+          <button>
+            <MdKeyboardArrowLeft className="btn" />
+          </button>
+          <button>
+            <MdKeyboardArrowRight className="btn" />
+          </button>
+        </div>
+      </div>
+    </S.Section>
+  );
+};
+
+export default SectionCate;
