+++ styles/globals.css
@@ -1,14 +1,10 @@
-
-@import './reset.css';
-
-
 html,
 body {
   padding: 0;
   margin: 0;
 
   font-weight: 100;
-  font-family: 'HallymGothic-Regular',-apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+  font-family: 'HallymGothic-Regular', -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 
 
@@ -25,8 +21,7 @@ a {
 
 @font-face {
   font-family: 'HallymGothic-Regular';
-  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2204@1.0/HallymGothic-Regular.woff2')
-    format('woff2');
+  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2204@1.0/HallymGothic-Regular.woff2') format('woff2');
   font-weight: 400;
   font-style: normal;
 }
@@ -36,6 +31,7 @@ a {
   html {
     color-scheme: dark;
   }
+
   body {
     color: white;
     background: black;
